requirements.txt
azure-storage-file-datalake==12.20.0
azure-ai-formrecognizer==3.3.3
python-dotenv==1.1.0
streamlit==1.45.1
pandas==2.2.3
openai==1.54.3